Break out all of the classes to the places they should go.
