+++ src/EventStores/SimpleCqrs.EventStore.SqlServer/ISqlStatementRunner.cs
@@ -1,0 +1,7 @@
+﻿namespace SimpleCqrs.EventStore.SqlServer
+{
+    public interface ISqlStatementRunner
+    {
+        void RunThisSql(string sqlStatement);
+    }
+}
