+++ src/EventStores/SimpleCqrs.EventStore.SqlServer/SimpleCqrs.EventStore.SqlServer.csproj
@@ -45,6 +45,7 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="IDomainEventSerializer.cs" />
+    <Compile Include="ISqlStatementRunner.cs" />
     <Compile Include="Serializers\JsonDomainEventSerializer.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="SqlServerConfiguration.cs" />
