+++ src/Tests/EventStores/SimpleCqrs.EventStore.SqlServer.Tests/SomethingHappenedEvent.cs
@@ -1,0 +1,8 @@
+﻿using SimpleCqrs.Eventing;
+
+namespace SimpleCqrs.EventStore.SqlServer.Tests
+{
+    public class SomethingHappenedEvent : DomainEvent
+    {
+    }
+}
