+++ src/Tests/EventStores/SimpleCqrs.EventStore.SqlServer.Tests/TestingRuntime.cs
@@ -1,0 +1,8 @@
+﻿using SimpleCqrs.Unity;
+
+namespace SimpleCqrs.EventStore.SqlServer.Tests
+{
+    public class TestingRuntime : SimpleCqrsRuntime<UnityServiceLocator>
+    {
+    }
+}
